libs: Use consolidated `aio.core`
