+++ aio.api.github/VERSION
@@ -1,1 +1,1 @@
-0.0.3-dev
+0.0.3
