+++ aio.api.github/aio/api/github/abstract/iterator.py
@@ -6,7 +6,7 @@ import gidgethub.sansio
 
 import abstracts
 
-from aio.functional import async_property
+from aio.core.functional import async_property
 
 
 class AGithubIterator(metaclass=abstracts.Abstraction):
