+++ aio.api.github/setup.cfg
@@ -29,7 +29,7 @@ py_modules = aio.api.github
 packages = find_namespace:
 install_requires =
     abstracts>=0.0.12
-    aio.functional>=0.0.10
+    aio.core>=0.2.0
     gidgethub
     packaging
