+++ envoy.abstract.command/envoy/abstract/command/BUILD
@@ -3,6 +3,5 @@ pytooling_library(
     "envoy.abstract.command",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
     ],
 )
