+++ envoy.code_format.python_check/VERSION
@@ -1,1 +1,1 @@
-0.0.5-dev
+0.0.5
