+++ envoy.code_format.python_check/envoy/code_format/python_check/BUILD
@@ -3,8 +3,7 @@ pytooling_library(
     "envoy.code_format.python_check",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.subprocess",
-        "//deps:aio.tasks",
+        "//deps:aio.core",
         "//deps:envoy.base.checker",
         "//deps:envoy.base.utils",
     ],
