+++ envoy.code_format.python_check/setup.cfg
@@ -29,8 +29,7 @@ py_modules = envoy.code_format.python_check
 packages = find_namespace:
 install_requires =
     abstracts>=0.0.12
-    aio.subprocess>=0.0.4
-    aio.tasks>=0.0.4
+    aio.core>=0.2.0
     envoy.base.checker>=0.1.0
     envoy.base.utils>=0.0.13
     flake8
