+++ envoy.dependency.cve_scan/VERSION
@@ -1,1 +1,1 @@
-0.0.2-dev
+0.0.2
