+++ envoy.dependency.cve_scan/setup.cfg
@@ -29,8 +29,7 @@ py_modules = envoy.dependency.cve_scan
 packages = find_namespace:
 install_requires =
     abstracts>=0.0.12
-    aio.functional>=0.0.10
-    aio.tasks>=0.0.4
+    aio.core>=0.2.0
     aiohttp
     envoy.base.checker>=0.0.3
     envoy.base.utils>=0.0.3
