+++ envoy.distribution.release/tests/BUILD
@@ -3,7 +3,7 @@ pytooling_tests(
     "envoy.distribution.release",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
+        "//deps:aio.core",
         "//deps:envoy.github.abstract",
         "//deps:envoy.github.release",
         "//deps:pytest-asyncio",
