+++ envoy.distribution.repo/envoy/distribution/repo/BUILD
@@ -3,8 +3,7 @@ pytooling_library(
     "envoy.distribution.repo",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
-        "//deps:aio.subprocess",
+        "//deps:aio.core",
         "//deps:envoy.base.runner",
         "//deps:envoy.base.utils",
     ],
