+++ envoy.distribution.repo/setup.cfg
@@ -29,8 +29,7 @@ py_modules = envoy.distribution.repo
 packages = find_namespace:
 install_requires =
     abstracts>=0.0.12
-    aio.functional>=0.0.10
-    aio.subprocess>=0.0.4
+    aio.core>=0.2.0
     envoy.base.runner>=0.1.0
     envoy.base.utils>=0.0.13
     envoy.abstract.command>=0.0.3
