+++ envoy.distribution.repo/tests/BUILD
@@ -3,7 +3,7 @@ pytooling_tests(
     "envoy.distribution.repo",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
+        "//deps:aio.core",
         "//deps:envoy.base.runner",
         "//deps:envoy.base.utils",
         "//deps:pytest-asyncio",
