+++ envoy.github.abstract/VERSION
@@ -1,1 +1,1 @@
-0.0.17-dev
+0.0.17
