+++ envoy.github.abstract/envoy/github/abstract/BUILD
@@ -3,8 +3,7 @@ pytooling_library(
     "envoy.github.abstract",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
-        "//deps:aio.tasks",
+        "//deps:aio.core",
         "//deps:envoy.abstract.command",
         "//deps:envoy.base.runner",
     ],
