+++ envoy.github.abstract/envoy/github/abstract/manager.py
@@ -14,7 +14,7 @@ import gidgethub.abc
 
 import abstracts
 
-from aio.functional import async_property
+from aio.core.functional import async_property
 
 from .release import AGithubRelease
