+++ envoy.github.abstract/envoy/github/abstract/release.py
@@ -10,7 +10,7 @@ import gidgethub.abc
 
 import abstracts
 
-from aio.functional import async_property
+from aio.core.functional import async_property
 
 from .assets import (
     AGithubReleaseAssetsFetcher, AGithubReleaseAssetsPusher)
