+++ envoy.github.abstract/setup.cfg
@@ -29,8 +29,7 @@ py_modules = envoy.github.abstract
 packages = find_namespace:
 install_requires =
     abstracts>=0.0.12
-    aio.functional>=0.0.10
-    aio.tasks>=0.0.4
+    aio.core>=0.2.0
     aiohttp
     envoy.base.runner>=0.1.0
     gidgethub
