+++ envoy.github.abstract/tests/BUILD
@@ -3,8 +3,7 @@ pytooling_tests(
     "envoy.github.abstract",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
-        "//deps:aio.tasks",
+        "//deps:aio.core",
         "//deps:envoy.abstract.command",
         "//deps:pytest-asyncio",
     ],
