+++ envoy.github.abstract/tests/test_manager.py
@@ -7,7 +7,7 @@ import packaging.version
 
 import abstracts
 
-from aio.functional import async_property
+from aio.core.functional import async_property
 from envoy.github.abstract import manager
