+++ envoy.github.release/envoy/github/release/BUILD
@@ -3,9 +3,7 @@ pytooling_library(
     "envoy.github.release",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.functional",
-        "//deps:aio.stream",
-        "//deps:aio.tasks",
+        "//deps:aio.core",
         "//deps:envoy.github.abstract",
         "//deps:envoy.base.utils",
     ],
