+++ envoy.github.release/setup.cfg
@@ -29,9 +29,7 @@ py_modules = envoy.github.release
 packages = find_namespace:
 install_requires =
     abstracts>=0.0.12
-    aio.functional>=0.0.10
-    aio.stream>=0.0.2
-    aio.tasks>=0.0.4
+    aio.core>=0.2.0
     aiohttp
     envoy.base.utils>=0.0.13
     envoy.github.abstract>=0.0.16
