+++ envoy.github.release/tests/BUILD
@@ -2,8 +2,7 @@
 pytooling_tests(
     "envoy.github.release",
     dependencies=[
-        "//deps:aio.functional",
-        "//deps:aio.tasks",
+        "//deps:aio.core",
         "//deps:envoy.github.abstract",
         "//deps:pytest-asyncio",
     ],
